Implementa sistema completo de cálculo de empacotamento e frete
- Reescreve edge function melhor-envio-quote com algoritmo de packing:
  classifica produtos por tipo (calçado/vestuário/acessório), cria blocos
  de ziplock e caixas rígidas, testa 4 estratégias de empacotamento com
  rotações e envia dimensões reais ao Melhor Envio
- Corrige parseFloat nos campos de dimensão do ProductFormModal para
  preservar valores decimais (ex: 21.5 cm)

+++ src/components/admin/ProductFormModal.tsx
@@ -290,9 +290,9 @@ export default function ProductFormModal({
         free_shipping: form.free_shipping,
         show_size_finder: form.show_size_finder,
         weight_g: form.weight_g ? parseInt(form.weight_g) : null,
-        pkg_height_cm: form.pkg_height_cm ? parseInt(form.pkg_height_cm) : null,
-        pkg_width_cm: form.pkg_width_cm ? parseInt(form.pkg_width_cm) : null,
-        pkg_length_cm: form.pkg_length_cm ? parseInt(form.pkg_length_cm) : null,
+        pkg_height_cm: form.pkg_height_cm ? parseFloat(form.pkg_height_cm) : null,
+        pkg_width_cm: form.pkg_width_cm ? parseFloat(form.pkg_width_cm) : null,
+        pkg_length_cm: form.pkg_length_cm ? parseFloat(form.pkg_length_cm) : null,
       };
 
       if (editingProduct) {
